reorganized layout to add utility folder
