+++ backend/src/main/java/com/abdi/cardiscover/controller/Brand.java
@@ -24,9 +24,9 @@ import com.abdi.cardiscover.repository.CarRepository;
 import com.abdi.cardiscover.repository.LocationRepository;
 import com.abdi.cardiscover.repository.ReservationRepository;
 import com.abdi.cardiscover.repository.SupplierRepository;
-import com.abdi.cardiscover.requestbody.CarRequestBody;
-import com.abdi.cardiscover.requestbody.ReservationRequestBody;
 import com.abdi.cardiscover.service.CleanCarData;
+import com.abdi.cardiscover.utility.requestbody.CarRequestBody;
+import com.abdi.cardiscover.utility.requestbody.ReservationRequestBody;
 
 
 @Component
