+++ backend/src/main/java/com/abdi/cardiscover/utility/requestbody/CarRequestBody.java
@@ -1,4 +1,4 @@
-package com.abdi.cardiscover.requestbody;
+package com.abdi.cardiscover.utility.requestbody;
 
 import java.math.BigDecimal;
 import java.util.GregorianCalendar;
