+++ frontend/src/index.html
@@ -5,7 +5,7 @@
   <title>Car Discover</title>
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <!-- <link rel="shortcut icon" type="image/x-icon" href="https://abdinasirnoor.com/static/landingpage/imgs/logo.ico" > -->
+  <link rel="shortcut icon" type="image/x-icon" href="https://abdinasirnoor.com/static/landingpage/imgs/logo.ico" >
   <link rel="preconnect" href="https://fonts.gstatic.com">
   <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap" rel="stylesheet">
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
